requirements.txt
spacy>=2.0.0
bigartm==0.9.0
